requirements.txt
aiofiles==23.2.1
alabaster==0.7.13
altair==5.1.1
annotated-types==0.5.0
anyio==3.7.1
appdirs==1.4.4
asttokens==2.0.5
attrs==23.1.0
autopep8==2.0.4
Babel==2.13.0
backcall==0.2.0
brotlipy==0.7.0
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==2.0.4
click==8.1.7
comm==0.1.2
contourpy==1.3.0
cryptography==39.0.1
cycler==0.12.1
debugpy==1.5.1
decorator==5.1.1
dgl==1.0.2+cu117
dglgo==0.0.2
docker-pycreds==0.4.0
docutils==0.20.1
einops==0.8.0
exceptiongroup==1.1.3
executing==0.8.3
fastapi==0.103.1
fastjsonschema==2.19.0
ffmpy==0.3.1
filelock==3.9.0
fonttools==4.54.1
fsspec==2024.10.0
gitdb==4.0.11
GitPython==3.1.43
gmpy2==2.1.2
gradio==3.42.0
gradio_client==0.5.0
h11==0.14.0
h5py==3.8.0
httpcore==0.17.3
httpx==0.24.1
huggingface-hub==0.26.2
idna==3.4
imageio==2.36.0
imagesize==1.4.1
imbalanced-learn==0.10.1
imblearn==0.0
importlib-resources==6.0.1
ipykernel==6.19.2
ipython==8.12.0
isort==5.12.0
jedi==0.18.1
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
jupyter_client==8.1.0
jupyter_core==5.3.0
kaleido==0.2.1
kiwisolver==1.4.7
kornia==0.6.12
lazy_loader==0.4
littleutils==0.2.2
MarkupSafe==2.1.1
matplotlib==3.9.2
matplotlib-inline==0.1.6
mkl-fft==1.3.6
mkl-random==1.2.2
mkl-service==2.4.0
mpmath==1.3.0
nbformat==5.9.2
nest-asyncio==1.5.6
networkx==2.8.4
numpy==1.24.3
numpydoc==1.6.0
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.77
nvidia-nvtx-cu12==12.1.105
ogb==1.3.6
opencv-python==4.10.0.84
orjson==3.9.5
outdated==0.2.2
packaging==23.0
pandas==2.0.1
parso==0.8.3
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
pip==23.0.1
platformdirs==2.5.2
plotly==5.18.0
point-transformer-pytorch==0.1.5
pooch==1.4.0
prompt-toolkit==3.0.36
protobuf==5.28.3
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.11.1
pycparser==2.21
pydantic==2.3.0
pydantic_core==2.6.3
pydub==0.25.1
Pygments==2.15.1
pyOpenSSL==23.0.0
pyparsing==3.2.0
PySocks==1.7.1
python-dateutil==2.8.2
python-multipart==0.0.6
pytz==2023.3
PyWavelets==1.4.1
PyYAML==6.0
pyzmq==25.0.2
rdkit-pypi==2022.9.5
referencing==0.30.2
requests==2.29.0
rpds-py==0.10.0
ruamel.yaml==0.17.39
ruamel.yaml.clib==0.2.8
safetensors==0.4.5
scikit-image==0.24.0
scikit-learn==1.2.2
scipy==1.10.1
semantic-version==2.10.0
sentry-sdk==2.18.0
setproctitle==1.3.3
setuptools==67.8.0
SimpleITK==2.4.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
snowballstemmer==2.2.0
Sphinx==7.2.6
sphinxcontrib-applehelp==1.0.7
sphinxcontrib-devhelp==1.0.5
sphinxcontrib-htmlhelp==2.0.4
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.6
sphinxcontrib-serializinghtml==1.1.9
stack-data==0.2.0
starlette==0.27.0
sympy==1.11.1
tabulate==0.9.0
tenacity==8.2.3
thop==0.1.1.post2209072238
threadpoolctl==3.1.0
tifffile==2024.9.20
timm==1.0.11
tomli==2.0.1
toolz==0.12.0
torch==2.3.1
torchaudio==2.3.1
torchvision==0.18.1
tornado==6.2
tqdm==4.65.0
traitlets==5.7.1
triton==2.3.1
typer==0.9.0
typing_extensions==4.11.0
tzdata==2023.3
urllib3==1.26.15
uvicorn==0.23.2
wandb==0.18.6
wcwidth==0.2.5
websockets==11.0.3
wheel==0.38.4
